create sidepanel for properties source
